% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drought_index_future.r
\name{drought_index_future}
\alias{drought_index_future}
\title{Drought Index For Stations for future projected rainfall}
\usage{
drought_index_future(data, years, droughtThreshold = 0.375)
}
\arguments{
\item{data}{a dataframe with date (future), year (future), month, rain (past) and rain_projected}

\item{years}{the number of years in the time series}

\item{droughtThreshold}{the level of dryness below which a drought begins}
}
\value{
dataframe with droughtIndices
}
\description{
Drought Index For Stations for future projected rainfall
}

