% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drought_index_grids.r
\name{drought_index_grids}
\alias{drought_index_grids}
\title{drought index using grids}
\usage{
drought_index_grids(rasterbrick, startyear, endyear, droughtThreshold = 0.375)
}
\arguments{
\item{rasterbrick}{a stack of grids}

\item{startyear}{the start year}

\item{endyear}{the end year}

\item{droughtThreshold}{the level of dryness}
}
\description{
drought index using grids
}

