\name{drought_index_stations}
\alias{drought_index_stations}
\title{Drought Index For Stations}
\usage{
  drought_index_stations(data, years,
    droughtThreshold = 0.375)
}
\arguments{
  \item{data}{a dataframe with date, year month and rain}
  \item{years}{the number of years in the time series}
  \item{droughtThreshold}{the level of dryness below which a drought begins}
}
\description{
  Drought Index For Stations
}

