
\name{HutchinsonDroughtIndex-package}
\alias{HutchinsonDroughtIndex-package}
\alias{HutchinsonDroughtIndex}
\docType{package}
\title{
Hutchinson Drought Index
}
\description{
Climatic Drought
}
\details{
\tabular{ll}{
Package: \tab HutchinsonDroughtIndex\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-14\cr
License: \tab GPL2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{


Maintainer: Who to complain to <ivan.hanigan@gmail.com>

}
\references{

}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
