
\name{HutchinsonDroughtIndex-package}
\alias{HutchinsonDroughtIndex-package}
\alias{HutchinsonDroughtIndex}
\docType{package}
\title{
Hutchinson Drought Index
}
\description{
Climatic Drought
}
\details{
\tabular{ll}{
Package: \tab HutchinsonDroughtIndex\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-11-06\cr
License: \tab GPL2\cr
}
The package is designed to take a timeseries for a single location such as region or weather station and compute the drought index.  Dev work on grid version.
}
\author{


Maintainer: <ivan.hanigan@gmail.com>

}
\references{

Smith, D. I, Hutchinson, M. F, & McArthur, R. J. (1992) Climatic and
Agricultural Drought: Payments and Policy. (Centre for Resource and
Environmental Studies, Australian National University, Canberra,
Australia). http://fennerschool-research.anu.edu.au/spatio-temporal/publications/cres_paper1992.pdf

Hanigan, IC. 2012. The Hutchinson Drought Index Algorithm [Computer
Software].  https://github.com/ivanhanigan/HutchinsonDroughtIndex

}
